Update WebData.py
Minor tweak to the "addresses" query to be more explicit. It does not change the functionality of the query. But preventative in the event the referenced fields later exists in one of the other referenced tables. By being explicit, it ensures that should a future update result in a name conflict from another table, the query will continue to work correctly.